requirements.txt
numpy==1.16.1
scipy==1.2.0
tensorflow==1.7.0
moviepy==1.0.1
Pillow==6.2.1